Fix incorrectly placed APIs.
